Index: Refactor index.js and separate app into app.js to allow for unit testing
